Grouped update socketIO event handlers in namespace app.blueprints.socket_io_event_handlers
