+++ source/app/blueprints/socket_io_event_handlers/case_event_handlers.py
@@ -56,3 +56,10 @@ def get_message(data):
     room = data['channel']
     join_room(room=room)
     emit('join', {'message': f"{current_user.user} just joined"}, room=room)
+
+
+@socket_io.on('join-case-obj-notif')
+@ac_socket_requires(CaseAccessLevel.full_access)
+def socket_join_case_obj_notif(data):
+    room = data['channel']
+    join_room(room=room)
